+++ src/ros/rover/arm/arm_kinematics/arm_kinematics/include/arm_kinematics/forward/utilities/analysis_tree.hpp
@@ -42,7 +42,7 @@ public:
    */
   struct JointDescription {
     JointType type = JointType::CONTINUOUS;
-    Eigen::Vector3d axis = Eigen::Vector3d::Zero();
+    Eigen::Vector3f axis = Eigen::Vector3f::Zero();
 
     JointDescription() = default;
     explicit JointDescription(const urdf::Joint & joint)
@@ -74,7 +74,7 @@ public:
     /// The type represented by this joint. Fixed if std::nullopt
     JointDescription joint{};
     /// Defines the reference frame of the link relative to the parent
-    Eigen::Isometry3d origin = Eigen::Isometry3d::Identity();
+    Eigen::Isometry3f origin = Eigen::Isometry3f::Identity();
 
     /// The IDs of all other frames that actuate relative to this link. All elements of this set should be larger than
     /// this link's id.
@@ -101,7 +101,7 @@ public:
     /// The ID of the link this is relative
     size_t parent = 0;
     /// The transform from the parent link to this frame
-    Eigen::Isometry3d origin;
+    Eigen::Isometry3f origin;
   };
 
   explicit AnalysisTree() {
@@ -133,7 +133,7 @@ public:
     const FrameDefinitions & definitions);
 
   Frame to_frame(const urdf::LinkConstSharedPtr & link) {
-    auto origin = Eigen::Isometry3d::Identity();
+    auto origin = Eigen::Isometry3f::Identity();
     const auto joint = find_next_non_fixed_joint(link.get(), origin);
 
     return {
@@ -146,7 +146,7 @@ public:
     const std::string & name,
     const size_t parent_id,
     const JointDescription & joint_description,
-    const Eigen::Isometry3d & origin)
+    const Eigen::Isometry3f & origin)
   {
     auto & parent = joints_[parent_id];
 
@@ -166,7 +166,7 @@ public:
   size_t add_frame(
     const std::string & name,
     const size_t parent_id,
-    const Eigen::Isometry3d & origin)
+    const Eigen::Isometry3f & origin)
   {
     const size_t id = frames_.add(name, {
       parent_id,
@@ -280,7 +280,7 @@ private:
    * nullopt is returned.
    * \return pointer to the joint's child link or nullptr if there is no joint, and this is relative to the root
    */
-  static urdf::Link const * find_next_non_fixed_joint(urdf::Link const * current, Eigen::Isometry3d & accumulator) {
+  static urdf::Link const * find_next_non_fixed_joint(urdf::Link const * current, Eigen::Isometry3f & accumulator) {
     // I unwrapped an originally recursive function
     while (current && current->parent_joint)
     {
@@ -335,7 +335,7 @@ private:
 
   /// Links with a non-fixed parent joint from the URDF, named after the parent joint
   /// In my model, I treat these as the same thing. The link will have a frame in frames_ relative to the joint with an
-  /// Eigen::Isometry3d::Identity origin.
+  /// Eigen::Isometry3f::Identity origin.
   NameToVector<Joint> joints_{};
 
   /// Links from the URDF
